+++ app/controllers/categories_controller.rb
@@ -1,0 +1,4 @@
+class CategoriesController < ApplicationController
+  def infrastructure_and_telecoms
+  end
+end