+++ CHANGELOG.md
@@ -58,6 +58,7 @@ This file is influenced by http://keepachangelog.com/.
 - Ability to edit page order
 - Give user option to update page parent
 - Ministers index page
+- Infrastructure and Telecommunications category (static)
 
 ### Changed